+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - **BREAKING CHANGE** : Require puppetlabs/dsc compiled with SQLServerDSC = 10.0.0.0
 - **BREAKING CHANGE** : Changed $sqlservicecredential_username and  $sqlagentservicecredential_username format. User accounts now required **WITHOUT** Netbios Domain Name prefix.
 - **BREAKING CHANGE** : Removed $domainName class parameter. Used facts instead.
+- **BREAKING CHANGE** : Removed $domainNetbiosName class parameter. Used custom fact instead.
 
 ## Version 1.1.0
 - Initial release